requirements.txt
python-dotenv
numpy
tifffile
matplotlib
pandas
seaborn
careamics==0.1.0rc8
